Fix entry of signing certificate password (#200)
* Fix entry of signing certificate password

The bug appears to be that the Save sheet conflicted with
the Security framework's password dialog, causing the inputs
to be lost. I fixed the problem by waiting for the sheet to be
dismissed before prompting for the password.

For some reason, this bug only reproduces for me on my Touch Bar
equipped MacBook Pro. Exporting signed profiles works just fine on
my iMac, also running macOS Catalina.

* Add explanatory comment